Split bwa aln and bwa index into multiple
