+++ bwapy/libbwaaln_utils.c
@@ -2,8 +2,8 @@
 #include "bwt.h"
 #include "bwtaln.h"
 #include "kstring.h"
-#include "libbwapy_utils.h"
 #include "bwase.h"
+#include "libbwaaln_utils.h"
 
 
 void bwa_cal_pac_pos_with_bwt(const bntseq_t *bns, int n_seqs, bwa_seq_t *seqs, int max_mm, float fnr, bwt_t *bwt)
