+++ bwapy/libbwaaln_utils.h
